Use int as data type for Discrete space
Since in C++ we currently support both size_t and int, this creates problem due to swig type mapping from python (which does not have unsigned types)

+++ gym/gympp/include/gympp/Space.h
@@ -72,7 +72,7 @@ extern template class gympp::spaces::details::TBox<double>;
 class gympp::spaces::Discrete : public gympp::spaces::Space
 {
 public:
-    using Type = size_t;
+    using Type = int;
     using Shape = gympp::data::Shape;
 
     Discrete() = delete;
